Crate rust_regex_explanation_pwa

rust_regex_explanation_pwa

version: 2020.803.1014 date: 2020-08-03 authors: Luciano Bestia
Rust regex explanations in PWA

regex explanation

Regex is great. But it is much easier to write and understand with a little help of explanations.
Regex has many flavors with subtle differences. This PWA uses Rust Regex crate.
We will need a web file server because security does not allow loading modules from local file.
Install this basic one: cargo install basic-http-server
Run the server in a separate terminal so it can stay running all the time:
Go to the content folder:
cd rustprojects/rust_regex_explanation_pwa/web_server_folder/web_content_folder
basic-http-server
Open the browser on:
http://127.0.0.1:4000

Minimal example of Wasm/Webassembly with vanilla Html, Css and javascript

First decision - no frameworks. Just vanilla. Then no javascript.
Some basic html. Some basic css.
All the rest is in Rust with web-sys/wasm-bindgen for all the programming needs.
No other special requirements.

PWA

Progressive Web App need a manifest, a worker and a bunch of icons.
todo...

todo: store the last input locally todo: read a json from file or url with the trio(regex,substitution,test_string)

Modules

code_gen_mod

regex_explanation_mod

regex_method_mod

web_sys_mod
helper functions for web_sys, window, document, dom, console, local_storage, session_storage,...

Functions

__wasm_bindgen_generated_wasm_bindgen_start
To start the Wasm application, wasm_bindgen runs this functions

on_keyup
code of event handler

set_event_listener_regex_text_on_keyup
set event listener for the regex_text

set_event_listener_substitution_on_keyup
set event listener for the substitution

set_event_listener_test_string_on_keyup
set event listener for the test_string

wasm_bindgen_start
To start the Wasm application, wasm_bindgen runs this functions