Crate rust_regex_explanation_pwa

rust_regex_explanation_pwa

version: 2020.810.542 date: 2020-08-10 authors: Luciano Bestia
Rust regex explanations in PWA

Try it

https://bestia.dev/rust_regex_explanation_pwa/

Regex explanation and testing

Regex is great. But it is much easier to write and understand with a little help of explanations.
Regex has many flavors with subtle differences. This PWA uses Rust Regex crate.
It generates Rust code and shows that code results for the 6 important Regex methods.
Useful as learning tool and as development tool to edit/test/debug your regex first.

Minimal example of Wasm/Webassembly with vanilla Html, Css and Javascript

First decision - no frameworks. Just vanilla. Then no javascript.
Some basic html. Some basic css.
contenteditable - Every day something new. Modern browsers have a wysiwyg editor.
Just add the attribute contenteditable to a div or code.
All the rest is in Rust with web-sys/wasm-bindgen for all the programming needs.
No other special requirements.
Ok, I had to use the javascript library highlightjs to bring some colors to the code.
It is more like a coding tutorial for creating PWA in Rust.

PWA

I added the manifest, the worker and a bunch of icons.
Now it :

can work offline,

can install as an "app"

execution is sand-boxed in the browser

no way to bring some malware (if the browser in not compromised)

Change colors

Users can change the colors with the use of the Chrome extension User CSS:
https://chrome.google.com/webstore/detail/user-css/okpjlejfhacmgjkmknjhadmkdbcldfcb
Copy/paste and then edit the colors and watch changes alive:

:root {
    /* color palette */
    /* background color */
    --b_color_body: #24292E;
    --b_color_code: #1B1D23;
    --b_color_grid_header: #181A1F;
    --b_color_container: #333842;
    --b_color_button: dodgerblue;
    /* front color */
    --f_color_body: #dce0e9;
    --f_color_code: #78C379;
    --f_color_link: white;
    --f_color_05: yellow;
    --f_color_06: dark-white;
    --f_color_07: black;
    /* border color*/
    --brd_color_01: #000;
    --brd_color_02: #eaecef;
}

Modules

code_gen_mod

examples_mod

regex_explanation_mod

regex_method_mod

test_string_mod

web_sys_mod
helper functions for web_sys, window, document, dom, console, local_storage, session_storage,...

Macros

html_encoded_push
Push html encoded variables inside unencoded literal to HtmlEncoded. Similar use as format! macro. html_encoded_push!(html, template, param_1) html_encoded_push!(exp.explanation_all.,"{}", &exp.reg_str)

on_click
Simple macro to set listener of onclick events to an element_id. no args: on_click!(element_1_id, function_ident) no args: on_click!("example_email",example_email) 1 args: on_click!(element_1_id, function_ident, arg_1) 1 args: on_click!("example_email",example_email, "arg_1") 2 args: on_click!(element_1_id, function_ident, arg_1,arg_2) 2 args: on_click!("example_email",example_email, "arg_1","arg_2")

on_keyup
Simple macro to set listener of onkeyup events to an element_id. on_keyup!(element_id, function_ident) on_keyup!("regex_text", run_regex)

Functions

__wasm_bindgen_generated_wasm_bindgen_start
To start the Wasm application, wasm_bindgen runs this functions

code_gen_copy

code_gen_run_in_playground

display_block_2_and_scroll
make visible 2 elements and jump to the third element

display_none_2
make 2 invisible

run_regex
on keyup code

set_all_event_listeners

wasm_bindgen_start
To start the Wasm application, wasm_bindgen runs this functions